+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [#74](https://github.com/wallix/awless/issues/74): support of encrypted SSH keys for generation `awless create keypair encrypted=true` and in `awless ssh`.
 - Create/delete and list Lambda functions
 - Create/delete and list autoscaling launch configurations and autoscaling groups
+- Create/delete and list cloudwatch alarms. List cloudwatch metrics.
 - List EC2 images (AMIs) of which you are the owner: `awless list images`
 - Copy an EC2 image from a given region to the current region: `awless copy image name=... source-id=... source-region=...`
 - List your IAM access keys: `awless list accesskeys`